Update unit test structure
