+++ package.json
@@ -225,7 +225,7 @@
 		"watch": "tsc -watch -p ./",
 		"pretest": "npm run compile && npm run lint",
 		"lint": "eslint src --ext .ts",
-		"test": "node ./out/test/runTests.js",
+		"test": "node ./out/test/runTest.js",
 		"convert": "npx js-yaml syntaxes/6x09.tmLanguage.yaml > syntaxes/6x09.tmLanguage.json",
 		"update": "npm run convert"
 	},
