--- src/test/runTests.ts
@@ -1,28 +1,0 @@
-import * as path from 'path';
-
-import { runTests } from '@vscode/test-electron';
-
-async function main(): Promise<void> {
-  try {
-    // The folder containing the Extension Manifest package.json
-    // Passed to `--extensionDevelopmentPath`
-    const extensionDevelopmentPath = path.resolve(__dirname, '../../');
-
-    // The path to the extension test runner script
-    // Passed to --extensionTestsPath
-    const extensionTestsPath = path.resolve(__dirname, './index');
-
-    // Download VS Code, unzip it and run the integration test
-    await runTests({
-      extensionDevelopmentPath,
-      extensionTestsPath,
-      launchArgs: ['--disable-extensions'],
-    });
-  } catch (err) {
-    // tslint:disable-next-line: no-console
-    console.error('Failed to run tests');
-    process.exit(1);
-  }
-}
-
-main();
