+++ src/test/suite/data/hello-clean.asm
