+++ src/test/suite/data/hello-errors.asm
